New benchmark for randomized svd
Randomized svd is faster than BDCSVD/gesdd when we need about 1/3 of the singular values.

To reach truncation error 10^-5 we need about 2/3 of all singular values. With this fraction, the BDCSVD/gesdd routines are faster than randomized svd.

+++ source/math/fit.h
@@ -27,7 +27,8 @@ namespace fit {
             Eigen::VectorXd X, Y;
             template<typename TX, typename TY>
             log_stretched_functor(const TX &xvec, const TY &yvec) : Functor(3, static_cast<int>(xvec.size())) {
-                if(xvec.size() != yvec.size()) throw std::runtime_error("fit::log_stretched: xvec and yvec size mismatch");
+                if(static_cast<size_t>(xvec.size()) != static_cast<size_t>(yvec.size()))
+                    throw std::runtime_error("fit::log_stretched: xvec and yvec size mismatch");
                 X = Eigen::Map<const Eigen::VectorXd>(xvec.data(), static_cast<Eigen::Index>(xvec.size()));
                 Y = Eigen::Map<const Eigen::VectorXd>(yvec.data(), static_cast<Eigen::Index>(yvec.size()));
             }
